Reestructuracion archivos js indicadores
