+++ src/main/resources/static/js/indicator/chart-config.js
@@ -1,0 +1,10 @@
+export const PADDING_AXIS_TITLE = 12;
+export const FONT_SIZE_AXIS_TITLE = 15;
+
+Chart.defaults.color = 'black';
+Chart.defaults.font.family = 'Arial';
+Chart.defaults.plugins.legend.position = 'right';
+Chart.defaults.plugins.legend.display = true;
+Chart.defaults.plugins.title.font.size = 25;
+Chart.defaults.plugins.title.display = true;
+Chart.defaults.responsive = true;
