+++ src/main/resources/static/js/indicator/training/chart-teacher.js
@@ -1,164 +1,11 @@
-const ctx = document.getElementById('myChart').getContext('2d');
-var myChart;
-
-const PADDING_AXIS_TITLE = 12;
-const FONT_SIZE_AXIS_TITLE = 15;
-
-Chart.defaults.color = 'black';
-Chart.defaults.font.family = 'Arial';
-Chart.defaults.plugins.legend.position = 'right';
-Chart.defaults.plugins.legend.display = true;
-Chart.defaults.plugins.title.font.size = 25;
-Chart.defaults.plugins.title.display = true;
-Chart.defaults.responsive = true;
-
-//TODO: Descomponer este archivo
-//TODO: Actualizar titulos genero - año en indicadore srestantes de docentes
-
-function defineTitle(data){
-    if(data === undefined) return '';
-
-    var titleSuffix = (data.length >= 2) ? data[data.length - 1] + " - " + data[0] : data[0];
-    return titleSuffix || '';
-}
-
-export function participantInstitutionCounterChart (labels, datasets, title) {
-  if(myChart) { myChart.destroy(); }
-
-    myChart = new Chart(ctx, {
-            type: 'bar',
-            data: {
-                labels: labels,
-                datasets: datasets
-            },
-            options: {
-                scales: {
-                    x: {
-                        position: 'top',
-                        title: {
-                            display: true,
-                            text: 'Cantidad por región',
-                            align: 'center',
-                            font: { size: FONT_SIZE_AXIS_TITLE },
-                            padding: { bottom: PADDING_AXIS_TITLE }
-                        }
-                    },
-                    y: {
-                        title: {
-                            display: true,
-                            text: 'Regiones',
-                            align: 'center',
-                            font: { size: FONT_SIZE_AXIS_TITLE },
-                            padding: { bottom: PADDING_AXIS_TITLE }
-                        }
-                    }
-                },
-                indexAxis: 'y',
-                plugins:{
-                    tooltip:{
-                      callbacks:{
-                        title: function(data){
-                          return data[0].label;
-                        },
-                        label: function(data){
-                            let value = data.formattedValue;
-                            let label = data.dataset.label;
-                            return label + ": " + "cantidad " + value;
-                        },
-
-                      },
-                    },
-                    title: {
-                        text: 'Número de establecimientos que participan en PNLD ' + defineTitle(title),
-                    }
-                }
-            }
-        });
-
-    myChart.update();
-}
-
-export function firstTimeInstitutionPercentageChart (labels, datasets, title, dataList) {
-  if(myChart) { myChart.destroy(); }
-
-  myChart = new Chart(ctx, {
-      type: 'bar',
-      data: { labels: labels, datasets: datasets },
-      options: {
-          scales: {
-              x: {
-                  position: 'top',
-                  min: 0,
-                  max: 100,
-                  ticks:{
-                    callback: function(value, index, values){
-                        return value+"%";
-                    }
-                  },
-                  title: {
-                      display: true,
-                      text: 'Porcentaje participación por primera vez',
-                      align: 'center',
-                      font: { size: FONT_SIZE_AXIS_TITLE },
-                      padding: { bottom: PADDING_AXIS_TITLE },
-                  }
-              },
-              y: {
-                  title: {
-                      display: true,
-                      text: 'Regiones',
-                      align: 'center',
-                      font: { size: FONT_SIZE_AXIS_TITLE },
-                      padding: { bottom : PADDING_AXIS_TITLE }
-                  }
-              }
-          },
-          indexAxis: 'y',
-          plugins:{
-              tooltip:{
-                  callbacks:{
-                    title: function(data){
-                      return data[0].label;
-                    },
-                    label: function(data){
-                        let percentage = data.formattedValue;
-                        let year = data.dataset.label;
-                        return year + ": " + "Porcentaje " + percentage + "%";
-                    },
-                    afterLabel: function(data){
-                      let label = data.label;
-                      let regionData = dataList.find(element => element.regionName === label);
-                      let year = parseInt(data.dataset.label);
-                      let dataCounter = regionData.trainingIndicatorDataList.map(e => {
-                          if (e.year === year) {
-                            let dataCounter = {
-                                'total': e.institutionCounterPNLD,
-                                'firstTime': e.firstTimeInstitutionCounter,
-                            };
-                            return dataCounter;
-                          };
-                      }).filter(e => typeof e != 'undefined');
-                      let total = (typeof dataCounter[0] != 'undefined') ? dataCounter[0].total : 0;
-                      let firstTimeCounter = (typeof dataCounter[0] != 'undefined') ? dataCounter[0].firstTime : 0;
-                      return "Detalle: " + "Total "+ total + " - valor actual " + firstTimeCounter;
-                    },
-                  },
-              },
-              title: {
-                  text: 'Porcentaje de establecimientos que participan por primera vez ' + defineTitle(title),
-              }
-          }
-      }
-  });
-
-  myChart.update();
-}
+import {PADDING_AXIS_TITLE, FONT_SIZE_AXIS_TITLE} from '../chart-config.js';
+import {defineTitle, getNewContext} from '../utils.js';
 
 export function trainedTeacherCounterChart(labels, datasets, title, keyword){
 
-    if(myChart) { myChart.destroy(); }
+    var ctx = getNewContext();
 
-    myChart = new Chart(ctx, {
+    var myChart = new Chart(ctx, {
             type: 'bar',
             data: {
                 labels: labels,
@@ -213,9 +60,9 @@ export function trainedTeacherCounterChart(labels, datasets, title, keyword){
 
 export function teacherInPersonSessionPercentageChart(labels, datasets, title, keyword, dataList){
 
-  if(myChart) { myChart.destroy(); }
+  var ctx = getNewContext();
 
-  myChart = new Chart(ctx, {
+  var myChart = new Chart(ctx, {
       type: 'bar',
       data: { labels: labels, datasets: datasets },
       options: {
@@ -289,9 +136,9 @@ export function teacherInPersonSessionPercentageChart(labels, datasets, title, k
 }
 
 export function teacherPretestCompletedPercentageChart(labels, datasets, title, keyword, dataList){
-  if(myChart) { myChart.destroy(); }
+  var ctx = getNewContext();
 
-  myChart = new Chart(ctx, {
+  var myChart = new Chart(ctx, {
       type: 'bar',
       data: { labels: labels, datasets: datasets },
       options: {
@@ -365,9 +212,9 @@ export function teacherPretestCompletedPercentageChart(labels, datasets, title, 
 }
 
 export function teacherPostTestCompletedPercentageChart(labels, datasets, title, keyword, dataList){
-  if(myChart) { myChart.destroy(); }
+  var ctx = getNewContext();
 
-  myChart = new Chart(ctx, {
+  var myChart = new Chart(ctx, {
       type: 'bar',
       data: { labels: labels, datasets: datasets },
       options: {
@@ -441,9 +288,9 @@ export function teacherPostTestCompletedPercentageChart(labels, datasets, title,
 }
 
 export function teacherTrainingCompletedPercentageChart(labels, datasets, title, keyword, dataList){
-  if(myChart) { myChart.destroy(); }
+  var ctx = getNewContext();
 
-  myChart = new Chart(ctx, {
+  var myChart = new Chart(ctx, {
       type: 'bar',
       data: { labels: labels, datasets: datasets },
       options: {
